moved coming soon to home
